+++ communicate-on-pull-request-released/__tests__/main.test.ts
@@ -1,6 +1,13 @@
 const path = require('path');
 const nock = require('nock');
 
+const validScenarios = [
+  {
+    response: 'release.json',
+    event_name: 'release'
+  }
+];
+
 const invalidScenarios = [
   {
     response: 'pull-request-closed.json'
@@ -13,6 +20,47 @@ const invalidScenarios = [
     event_name: 'release'
   }
 ];
+
+describe('action test suite', () => {
+  for (const scenario of validScenarios) {
+    it(`It posts a comment on pull requests, referenced issues and update labels for (${scenario.response})`, async () => {
+      process.env['INPUT_REPO-TOKEN'] = 'token';
+      process.env['INPUT_PR-LABEL-TO-ADD'] = 'label-to-add';
+      process.env['INPUT_PR-LABEL-TO-REMOVE'] = 'label-to-remove';
+
+      process.env['GITHUB_REPOSITORY'] = 'foo/bar';
+      process.env['GITHUB_EVENT_NAME'] = scenario.event_name;
+      process.env['GITHUB_EVENT_PATH'] = path.join(
+        __dirname,
+        scenario.response
+      );
+
+      const api = nock('https://api.github.com')
+        .persist()
+        .post(
+          '/repos/foo/bar/pulls/999/reviews',
+          '{"body":"Congratulations! :tada: This was released as part of [_fastlane_ 2.134.1](https://github.com/Codertocat/Hello-World/runs/128620228) :rocket:","event":"COMMENT"}'
+        )
+        .reply(200)
+        .get('/repos/foo/bar/issues/999/labels')
+        .reply(200, JSON.parse('[]'))
+        .post('/repos/foo/bar/issues/999/labels', '{"labels":["label-to-add"]}')
+        .reply(200)
+        .get('/repos/foo/bar/pulls/999')
+        .reply(200, JSON.parse('{"body":"closes #10"}'))
+        .post(
+          '/repos/foo/bar/issues/10/comments',
+          '{"body":"The pull request #999 that closed this issue was merged and released as part of [_fastlane_ 2.134.1](https://github.com/Codertocat/Hello-World/runs/128620228) :rocket:\\nPlease let us know if the functionality works as expected as a reply here. If it does not, please open a new issue. Thanks!"}'
+        )
+        .reply(200);
+
+      const main = require('../src/main');
+      await main.run();
+
+      expect(api.isDone()).toBeTruthy();
+    });
+  }
+});
 
 describe('action test suite', () => {
   for (const scenario of invalidScenarios) {
@@ -31,8 +79,6 @@ describe('action test suite', () => {
       const api = nock('https://api.github.com')
         .persist()
         .post('/repos/foo/bar/issues/10/labels', '{"labels":["label-to-add"]}')
-        .reply(200)
-        .delete('repos/foo/bar/issues/20/labels/label-to-remove')
         .reply(200);
 
       const main = require('../src/main');
