+++ communicate-on-pull-request-released/__tests__/release.json
@@ -1,0 +1,8 @@
+{
+    "release": {
+        "tag_name": "2.134.1",
+        "body": "* [action] title (#999) via XYZ",
+        "html_url": "https://github.com/Codertocat/Hello-World/runs/128620228"
+    }, 
+    "action": "published"
+}
